% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbox_to_wkt.R
\name{bbox_to_wkt}
\alias{bbox_to_wkt}
\title{Convertir bbox a formato WKT optimizado para API específica}
\usage{
bbox_to_wkt(bbox, api_type = "gbif")
}
\arguments{
\item{bbox}{String "min_lon,min_lat,max_lon,max_lat" o vector numérico}

\item{api_type}{Tipo de API: "gbif", "obis", "inat", "idigbio"}
}
\value{
String WKT optimizado para la API específica
}
\description{
Convertir bbox a formato WKT optimizado para API específica
}
