% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_spatial_statistics.R
\name{calculate_spatial_statistics}
\alias{calculate_spatial_statistics}
\title{Calcular estadísticas espaciales}
\usage{
calculate_spatial_statistics(data)
}
\arguments{
\item{data}{Data frame con columnas lon y lat}
}
\value{
Lista con estadísticas espaciales
}
\description{
Genera resumen estadístico de la distribución espacial de los puntos
}
