% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_taxonomy.R
\name{check_taxonomy}
\alias{check_taxonomy}
\title{Check taxonomy using WORMS}
\usage{
check_taxonomy(thisspecies)
}
\arguments{
\item{thisspecies}{}
}
\value{
sp_data
}
\description{
Check taxonomy using WORMS
}
