% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_data.R
\name{combine_data}
\alias{combine_data}
\title{Combine data}
\usage{
combine_data(file.list, all_results)
}
\arguments{
\item{file.list}{WKT (Well-Known Text) string para la geometría de búsqueda}

\item{all_results}{Lista de configuración GBIF con parámetros de consulta}
}
\value{
Data frame con registros de GBIF formateados
}
\description{
Combine data
}
