% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_spatial_points.R
\name{create_spatial_points}
\alias{create_spatial_points}
\title{Crear objeto espacial sf desde coordenadas}
\usage{
create_spatial_points(data, crs = 4326)
}
\arguments{
\item{data}{Data frame con columnas lon y lat}

\item{crs}{Sistema de coordenadas (por defecto 4326)}
}
\value{
Objeto sf
}
\description{
Convierte un data frame con coordenadas a objeto sf
}
