% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/execute_all_queries.R
\name{execute_all_queries}
\alias{execute_all_queries}
\title{Ejecutar consultas para todas las bases de datos habilitadas}
\usage{
execute_all_queries(grid, config, log_function)
}
\arguments{
\item{grid}{Data frame con columnas box_id, bbox, wkt}

\item{config}{Lista de configuración completa}

\item{log_function}{Función para registrar mensajes}
}
\value{
Data frame con todos los resultados consolidados
}
\description{
Ejecutar consultas para todas las bases de datos habilitadas
}
