% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_points_in_polygon.R
\name{filter_points_in_polygon}
\alias{filter_points_in_polygon}
\title{Filtrar puntos dentro de polígono}
\usage{
filter_points_in_polygon(points_data, polygon)
}
\arguments{
\item{points_data}{Data frame con columnas lon y lat}

\item{polygon}{Objeto sf con polígono}
}
\value{
Data frame filtrado con puntos dentro del polígono
}
\description{
Selecciona solo los puntos que caen dentro de un polígono dado
}
