% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_biodiversity_data_gbif.R
\name{format_biodiversity_data_gbif}
\alias{format_biodiversity_data_gbif}
\title{Formatear datos de biodiversidad a estructura est치ndar}
\usage{
format_biodiversity_data_gbif(data, source, config)
}
\arguments{
\item{data}{Data frame con datos brutos de biodiversidad}

\item{source}{Nombre de la fuente de datos (ej: "GBIF", "OBIS")}

\item{config}{Nombre de la fuente de datos (ej: "GBIF", "OBIS")}
}
\value{
Data frame formateado y validado
}
\description{
Convierte datos de diferentes fuentes a un formato unificado con
validaci칩n de tipos y eliminaci칩n de registros inv치lidos
}
