% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_simple_bbox.R
\name{generate_simple_bbox}
\alias{generate_simple_bbox}
\title{Generar bounding box simple de polígono}
\usage{
generate_simple_bbox(polygon)
}
\arguments{
\item{polygon}{Objeto sf con geometría}
}
\value{
Data frame con una fila: box_id, bbox, wkt
}
\description{
Función auxiliar para generar un único bounding box cuando no se usa grid
}
