% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_grid_statistics.R
\name{get_grid_statistics}
\alias{get_grid_statistics}
\title{Obtener estadísticas del grid optimizado}
\usage{
get_grid_statistics(grid, original_polygon)
}
\arguments{
\item{grid}{Grid sf object}

\item{original_polygon}{Polígono original}
}
\value{
Lista con estadísticas del grid
}
\description{
Obtener estadísticas del grid optimizado
}
