% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_query_config.R
\name{load_query_config}
\alias{load_query_config}
\title{Cargar configuraci贸n desde archivo JSON}
\usage{
load_query_config(config_file)
}
\arguments{
\item{config_file}{Ruta al archivo JSON de configuraci贸n}
}
\value{
Lista con la configuraci贸n cargada
}
\description{
Cargar configuraci贸n desde archivo JSON
}
