% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_fn.R
\name{log_fn}
\alias{log_fn}
\title{Función de logging que escribe en archivo y consola}
\usage{
log_fn(msg, log_file, level = "INFO")
}
\arguments{
\item{msg}{message}

\item{level}{information}
}
\description{
Función de logging que escribe en archivo y consola
}
