% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_function.R
\name{log_function}
\alias{log_function}
\title{Wrapper para función de logging que escribe en archivo y consola}
\usage{
log_function(msg, log_file, level = "INFO")
}
\arguments{
\item{msg}{message}

\item{level}{information}
}
\description{
Wrapper para función de logging que escribe en archivo y consola
}
