% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/min_bounding_polygon.R
\name{min_bounding_polygon}
\alias{min_bounding_polygon}
\title{Get minimum bounding polygon}
\usage{
min_bounding_polygon(polygon)
}
\arguments{
\item{polygon}{}
}
\description{
https://johngodlee.xyz/posts/2021-11-14-minimum_rectangle/
}
