% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_gbif.R
\name{query_gbif}
\alias{query_gbif}
\title{Consultar GBIF (Global Biodiversity Information Facility)}
\usage{
query_gbif(grid_row, config, box_id, this_source)
}
\arguments{
\item{grid_row}{WKT (Well-Known Text) string para la geometría de búsqueda}

\item{config}{Lista de configuración GBIF con parámetros de consulta}

\item{box_id}{ID de la caja de búsqueda (para logging)}

\item{this_source}{database name}
}
\value{
Saves as cvs records found
}
\description{
Consultar GBIF (Global Biodiversity Information Facility)
}
