% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_inat.R
\name{query_inat}
\alias{query_inat}
\title{Consultar iNaturalist}
\usage{
query_inat(grid_row, config, box_id = 1, log_function)
}
\arguments{
\item{grid_row}{Fila del grid con información espacial completa}

\item{config}{Lista de configuración iNaturalist con parámetros de consulta}

\item{box_id}{ID de la caja de búsqueda (para logging)}

\item{log_function}{Función para registrar mensajes (opcional)}
}
\value{
Data frame con registros de iNaturalist formateados
}
\description{
Consultar iNaturalist
}
