% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_duplicates.R
\name{remove_duplicates}
\alias{remove_duplicates}
\title{Eliminar registros duplicados de biodiversidad}
\usage{
remove_duplicates(data, keep_first = TRUE)
}
\arguments{
\item{data}{Data frame con registros de biodiversidad}

\item{keep_first}{Lógico, mantener primera ocurrencia (TRUE) o última (FALSE)}
}
\value{
Data frame sin duplicados
}
\description{
Elimina duplicados basándose en combinación de especies, coordenadas, año y fuente
}
