% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_parallel_biodiversity_query.R
\name{run_parallel_biodiversity_query}
\alias{run_parallel_biodiversity_query}
\title{Ejecutar consulta paralela mejorada para todas las bases de datos}
\usage{
run_parallel_biodiversity_query(
  polygon_file = "shapefiles/study_zone.gpkg",
  config_file = "scripts/config/query_config_maxima.json"
)
}
\arguments{
\item{polygon_file}{Ruta al archivo del polígono de estudio}

\item{config_file}{Ruta al archivo de configuración}
}
\value{
Lista con resultados de la consulta
}
\description{
Ejecutar consulta paralela mejorada para todas las bases de datos
}
