% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplify_polygon.R
\name{simplify_polygon}
\alias{simplify_polygon}
\title{Simplificar geometría de polígono}
\usage{
simplify_polygon(polygon, tolerance = 0.01)
}
\arguments{
\item{polygon}{Objeto sf con geometría}

\item{tolerance}{Tolerancia para simplificación (grados)}
}
\value{
Objeto sf simplificado
}
\description{
Reduce la complejidad de un polígono manteniendo su forma general
}
