% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_coordinates.R
\name{validate_coordinates}
\alias{validate_coordinates}
\title{Validar coordenadas geográficas}
\usage{
validate_coordinates(lon, lat)
}
\arguments{
\item{lon}{Vector de longitudes}

\item{lat}{Vector de latitudes}
}
\value{
Vector lógico indicando coordenadas válidas
}
\description{
Verifica que las coordenadas estén dentro de rangos válidos
}
